requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
pkg-resources==0.0.0
PySocks==1.7.0
requests==2.22.0
urllib3==1.25.2
